Separated CartItem amd Product types
