+++ client/src/components/Product.tsx
@@ -12,7 +12,7 @@ const Product = ({product}: ProductProps) => {
         <p className="price">${product.price}</p>
         <p className="quantity">{product.quantity} left in stock</p>
         <div className="actions product-actions">
-          <button className="add-to-cart">Add to Cart</button>
+          <button className="add-to-cart" disabled={product.quantity === 0}>Add to Cart</button>
           <button className="edit">Edit</button>
         </div>
         <button className="delete-button"><span>X</span></button>
