+++ client/src/types/CartItem.tsx
@@ -1,0 +1,5 @@
+import type {Product} from "./Product";
+
+export interface CartItem extends Product{
+  productId: string;
+}
